chatbot/requirements.txt
langchain-google-genai
streamlit
python_dotenv
# google-generativeai